requirements.txt
PyYAML==5.3.1
PyJWT==1.7.1
passlib==1.7.2
Cerberus==1.3.2
bcrypt==3.1.7
boto3==1.14.44
aioboto3==8.2.0
aiofile==3.3.3
